+++ src/components/index.ts
@@ -3,3 +3,4 @@ export * from './PlayerView/pictureInPictureConfig';
 export * from './PlayerView/playerViewConfig';
 export * from './PlayerView/properties';
 export * from './PlayerView/playerViewRef';
+export * from './PlayerView/pictureInPictureAction';