+++ CHANGELOG.md
@@ -4,10 +4,9 @@
 
 ### Added
 
-- `PictureInPictureConfig.pictureInPictureActions` to configure Picture in Picture actions
+- `BasePlayerViewProps.pictureInPictureActions` to configure Picture in Picture actions
   - `PictureInPictureAction.TogglePlayback` to add a play/pause action
   - `PictureInPictureAction.Seek` to add seek forward and seek backward for 10 seconds actions
-- `PlayerViewRef.updatePictureInPictureActions` to update Picture in Picture actions dynamically
 
 ### Changed